+++ packages/rum/src/domain/record/observers.spec.ts
@@ -5,8 +5,8 @@ import type { RawRumEventCollectedData } from 'packages/rum-core/src/domain/life
 import { createNewEvent } from '../../../../core/test/specHelper'
 import { NodePrivacyLevel, PRIVACY_ATTR_NAME, PRIVACY_ATTR_VALUE_MASK_USER_INPUT } from '../../constants'
 import { RecordType } from '../../types'
-import type { FrustrationCallback, InputCallback } from './observer'
-import { initFrustrationObserver, initInputObserver } from './observer'
+import type { FrustrationCallback, InputCallback } from './observers'
+import { initFrustrationObserver, initInputObserver } from './observers'
 import { serializeDocument } from './serialize'
 
 describe('initInputObserver', () => {
